requirements.txt
Flask==2.0.2
werkzeug>=1,<2.1
Flask-Session>=0,<1
requests>=2,<3
msal>=1.7,<2
CacheLib
Flask-SQLAlchemy==2.5.1
PyMySQL
flask_sqlalchemy_session
objectpath
azure-iot-hub



#Flask>=2,<3
#flask-session~=0.3.2
#werkzeug>=1,<2
#flask-session>=0.3.2,<0.5
#requests>=2,<3
#msal>=1.7,<2

#Flask>=1,<2
#Flask-Session>=0,<1
#requests>=2,<3
#msal>=0,<2
